Podspec version update from 6.0 to 7.0
Update podspec version from 6.0 to 7.0.
Release version (tags) are mapped to a certain commit.
6.0 is mapped to 7fe656b

Cocoapods clones that commit in stead of c27f9e1 - the one that has the
latest podspec and the AES encryption support.

Updating the podspec version to 7.0 and creating a new release (7.0)
would allow CocoaPods to clone the correct version.